backend/main.py

from fastapi import FastAPI, UploadFile, File, HTTPException
from pydantic import BaseModel
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

app = FastAPI()

class Query(BaseModel):
    query: str

@app.get("/")
def read_root():
    return {"message": "Welcome to the InsightBot API"}

@app.post("/api/query")
def query(query: Query):
    # This is where the LangGraph agent logic will go
    return {"answer": "This is a placeholder answer.", "sources": [], "evaluation": {"score": 0, "rationale": ""}}

@app.post("/api/upload")
def upload_file(file: UploadFile = File(...)):
    # This is where the file upload and embedding logic will go
    return {"message": f"Successfully uploaded {file.filename}"}

@app.get("/api/queries")
def get_queries():
    # This is where the query history and evaluation scores will be retrieved
    return []

@app.post("/api/eval")
def eval_result():
    # This is where a result will be re-evaluated
    return {"message": "This is a placeholder for re-evaluation."}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
