backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
langchain==0.0.335
langchain-openai==0.0.2
langgraph==0.0.40
faiss-cpu==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.24.4
sentence-transformers==2.2.2
loguru==0.7.2
tenacity==8.2.3
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
mcp==1.0.0